+++ examples/derive_ref/flatten_hand_args.rs
@@ -1,5 +1,5 @@
 use clap::error::Error;
-use clap::{Arg, ArgMatches, Args, Command, FromArgMatches, Parser};
+use clap::{Arg, ArgAction, ArgMatches, Args, Command, FromArgMatches, Parser};
 
 #[derive(Debug)]
 struct CliArgs {
@@ -15,8 +15,8 @@ impl FromArgMatches for CliArgs {
     }
     fn from_arg_matches_mut(matches: &mut ArgMatches) -> Result<Self, Error> {
         Ok(Self {
-            foo: matches.is_present("foo"),
-            bar: matches.is_present("bar"),
+            foo: *matches.get_one::<bool>("foo").expect("defaulted by clap"),
+            bar: *matches.get_one::<bool>("bar").expect("defaulted by clap"),
             quuz: matches.remove_one::<String>("quuz"),
         })
     }
@@ -25,8 +25,8 @@ impl FromArgMatches for CliArgs {
         self.update_from_arg_matches_mut(&mut matches)
     }
     fn update_from_arg_matches_mut(&mut self, matches: &mut ArgMatches) -> Result<(), Error> {
-        self.foo |= matches.is_present("foo");
-        self.bar |= matches.is_present("bar");
+        self.foo |= *matches.get_one::<bool>("foo").expect("defaulted by clap");
+        self.bar |= *matches.get_one::<bool>("bar").expect("defaulted by clap");
         if let Some(quuz) = matches.remove_one::<String>("quuz") {
             self.quuz = Some(quuz);
         }
@@ -36,20 +36,40 @@ impl FromArgMatches for CliArgs {
 
 impl Args for CliArgs {
     fn augment_args(cmd: Command<'_>) -> Command<'_> {
-        cmd.arg(Arg::new("foo").short('f').long("foo"))
-            .arg(Arg::new("bar").short('b').long("bar"))
-            .arg(Arg::new("quuz").short('q').long("quuz").takes_value(true))
+        cmd.arg(
+            Arg::new("foo")
+                .short('f')
+                .long("foo")
+                .action(ArgAction::SetTrue),
+        )
+        .arg(
+            Arg::new("bar")
+                .short('b')
+                .long("bar")
+                .action(ArgAction::SetTrue),
+        )
+        .arg(Arg::new("quuz").short('q').long("quuz").takes_value(true))
     }
     fn augment_args_for_update(cmd: Command<'_>) -> Command<'_> {
-        cmd.arg(Arg::new("foo").short('f').long("foo"))
-            .arg(Arg::new("bar").short('b').long("bar"))
-            .arg(Arg::new("quuz").short('q').long("quuz").takes_value(true))
+        cmd.arg(
+            Arg::new("foo")
+                .short('f')
+                .long("foo")
+                .action(ArgAction::SetTrue),
+        )
+        .arg(
+            Arg::new("bar")
+                .short('b')
+                .long("bar")
+                .action(ArgAction::SetTrue),
+        )
+        .arg(Arg::new("quuz").short('q').long("quuz").takes_value(true))
     }
 }
 
 #[derive(Parser, Debug)]
 struct Cli {
-    #[clap(short, long)]
+    #[clap(short, long, action)]
     top_level: bool,
     #[clap(flatten)]
     more_args: CliArgs,