+++ tests/builder/utf8.rs
@@ -76,132 +76,13 @@ fn invalid_utf8_strict_option_long_equals() {
 }
 
 #[test]
-fn invalid_utf8_lossy_positional() {
-    #![allow(deprecated)]
-    let r = Command::new("bad_utf8")
-        .arg(Arg::new("arg").allow_invalid_utf8(true))
-        .try_get_matches_from(vec![OsString::from(""), OsString::from_vec(vec![0xe9])]);
-    assert!(r.is_ok(), "{}", r.unwrap_err());
-    let m = r.unwrap();
-    assert!(m.is_present("arg"));
-    assert_eq!(&*m.value_of_lossy("arg").unwrap(), "\u{FFFD}");
-}
-
-#[test]
-fn invalid_utf8_lossy_option_short_space() {
-    #![allow(deprecated)]
-    let r = Command::new("bad_utf8")
-        .arg(
-            Arg::new("arg")
-                .short('a')
-                .long("arg")
-                .takes_value(true)
-                .allow_invalid_utf8(true),
-        )
-        .try_get_matches_from(vec![
-            OsString::from(""),
-            OsString::from("-a"),
-            OsString::from_vec(vec![0xe9]),
-        ]);
-    assert!(r.is_ok(), "{}", r.unwrap_err());
-    let m = r.unwrap();
-    assert!(m.is_present("arg"));
-    assert_eq!(&*m.value_of_lossy("arg").unwrap(), "\u{FFFD}");
-}
-
-#[test]
-fn invalid_utf8_lossy_option_short_equals() {
-    #![allow(deprecated)]
-    let r = Command::new("bad_utf8")
-        .arg(
-            Arg::new("arg")
-                .short('a')
-                .long("arg")
-                .takes_value(true)
-                .allow_invalid_utf8(true),
-        )
-        .try_get_matches_from(vec![
-            OsString::from(""),
-            OsString::from_vec(vec![0x2d, 0x61, 0x3d, 0xe9]),
-        ]);
-    assert!(r.is_ok(), "{}", r.unwrap_err());
-    let m = r.unwrap();
-    assert!(m.is_present("arg"));
-    assert_eq!(&*m.value_of_lossy("arg").unwrap(), "\u{FFFD}");
-}
-
-#[test]
-fn invalid_utf8_lossy_option_short_no_space() {
-    #![allow(deprecated)]
-    let r = Command::new("bad_utf8")
-        .arg(
-            Arg::new("arg")
-                .short('a')
-                .long("arg")
-                .takes_value(true)
-                .allow_invalid_utf8(true),
-        )
-        .try_get_matches_from(vec![
-            OsString::from(""),
-            OsString::from_vec(vec![0x2d, 0x61, 0xe9]),
-        ]);
-    assert!(r.is_ok(), "{}", r.unwrap_err());
-    let m = r.unwrap();
-    assert!(m.is_present("arg"));
-    assert_eq!(&*m.value_of_lossy("arg").unwrap(), "\u{FFFD}");
-}
-
-#[test]
-fn invalid_utf8_lossy_option_long_space() {
-    #![allow(deprecated)]
-    let r = Command::new("bad_utf8")
-        .arg(
-            Arg::new("arg")
-                .short('a')
-                .long("arg")
-                .takes_value(true)
-                .allow_invalid_utf8(true),
-        )
-        .try_get_matches_from(vec![
-            OsString::from(""),
-            OsString::from("--arg"),
-            OsString::from_vec(vec![0xe9]),
-        ]);
-    assert!(r.is_ok(), "{}", r.unwrap_err());
-    let m = r.unwrap();
-    assert!(m.is_present("arg"));
-    assert_eq!(&*m.value_of_lossy("arg").unwrap(), "\u{FFFD}");
-}
-
-#[test]
-fn invalid_utf8_lossy_option_long_equals() {
-    #![allow(deprecated)]
-    let r = Command::new("bad_utf8")
-        .arg(
-            Arg::new("arg")
-                .short('a')
-                .long("arg")
-                .takes_value(true)
-                .allow_invalid_utf8(true),
-        )
-        .try_get_matches_from(vec![
-            OsString::from(""),
-            OsString::from_vec(vec![0x2d, 0x2d, 0x61, 0x72, 0x67, 0x3d, 0xe9]),
-        ]);
-    assert!(r.is_ok(), "{}", r.unwrap_err());
-    let m = r.unwrap();
-    assert!(m.is_present("arg"));
-    assert_eq!(&*m.value_of_lossy("arg").unwrap(), "\u{FFFD}");
-}
-
-#[test]
 fn invalid_utf8_positional() {
     let r = Command::new("bad_utf8")
         .arg(Arg::new("arg").value_parser(value_parser!(OsString)))
         .try_get_matches_from(vec![OsString::from(""), OsString::from_vec(vec![0xe9])]);
     assert!(r.is_ok(), "{}", r.unwrap_err());
     let m = r.unwrap();
-    assert!(m.is_present("arg"));
+    assert!(m.contains_id("arg"));
     assert_eq!(
         &*m.get_one::<OsString>("arg").unwrap(),
         &*OsString::from_vec(vec![0xe9])
@@ -225,7 +106,7 @@ fn invalid_utf8_option_short_space() {
         ]);
     assert!(r.is_ok(), "{}", r.unwrap_err());
     let m = r.unwrap();
-    assert!(m.is_present("arg"));
+    assert!(m.contains_id("arg"));
     assert_eq!(
         &*m.get_one::<OsString>("arg").unwrap(),
         &*OsString::from_vec(vec![0xe9])
@@ -248,7 +129,7 @@ fn invalid_utf8_option_short_equals() {
         ]);
     assert!(r.is_ok(), "{}", r.unwrap_err());
     let m = r.unwrap();
-    assert!(m.is_present("arg"));
+    assert!(m.contains_id("arg"));
     assert_eq!(
         &*m.get_one::<OsString>("arg").unwrap(),
         &*OsString::from_vec(vec![0xe9])
@@ -271,7 +152,7 @@ fn invalid_utf8_option_short_no_space() {
         ]);
     assert!(r.is_ok(), "{}", r.unwrap_err());
     let m = r.unwrap();
-    assert!(m.is_present("arg"));
+    assert!(m.contains_id("arg"));
     assert_eq!(
         &*m.get_one::<OsString>("arg").unwrap(),
         &*OsString::from_vec(vec![0xe9])
@@ -295,7 +176,7 @@ fn invalid_utf8_option_long_space() {
         ]);
     assert!(r.is_ok(), "{}", r.unwrap_err());
     let m = r.unwrap();
-    assert!(m.is_present("arg"));
+    assert!(m.contains_id("arg"));
     assert_eq!(
         &*m.get_one::<OsString>("arg").unwrap(),
         &*OsString::from_vec(vec![0xe9])
@@ -318,7 +199,7 @@ fn invalid_utf8_option_long_equals() {
         ]);
     assert!(r.is_ok(), "{}", r.unwrap_err());
     let m = r.unwrap();
-    assert!(m.is_present("arg"));
+    assert!(m.contains_id("arg"));
     assert_eq!(
         &*m.get_one::<OsString>("arg").unwrap(),
         &*OsString::from_vec(vec![0xe9])